Cleaning up and simplifying CMake setup for Python.
This does not rely on the CMake/Swig stuff but calls swig and
python setuptools directly. This way, it is not necessary to
figure out the Python and Numpy libraries and include dirs in
CMake, because this does not work correctly.

The Python interface is now switched off be default and can
be enabled with the --python option to the setup script.

I still kept the numpy detection in CMake because Numpy is
required and CMake should fail if it is not there.