Add pytest dependency
`ctest` proceeds to test the Python bindings via the `pytest`, which is not a built-in Python package.
Thus, `ctest` fails if `pytest` is not installed. The source of failure is only visible when `ctest -V` is invoked.

It might be a good idea to document an optional `pytest` dependency here and/or install it automatically (as `pybind11`).